requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Authlib==1.2.0
blinker==1.6.2
boto3==1.26.120
botocore==1.29.120
cachelib==0.9.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
colorlover==0.3.0
cryptography==39.0.0
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
exceptiongroup==1.1.0
executing==1.2.0
Flask==2.2.2
Flask-Caching==1.10.1
Flask-Login==0.6.2
frozenlist==1.3.3
gics==0.1a6
icecream==2.1.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
libsass==0.22.0
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pandas==1.3.5
patsy==0.5.3
pg8000==1.29.4
plotly==5.13.0
pluggy==1.0.0
psycopg2-binary==2.9.6
pyarrow==11.0.0
pycares==4.3.0
pycparser==2.21
Pygments==2.15.0
pytest==7.3.1
python-dateutil==2.8.2
pytz==2022.7.1
regex==2023.3.23
requests==2.28.2
s3transfer==0.6.0
scramp==1.4.4
six==1.16.0
SQLAlchemy==1.4.47
tabulate==0.9.0
tenacity==8.1.0
testing.common.database==2.0.3
testing.postgresql==1.3.0
threadpoolctl==3.1.0
tomli==2.0.1
urllib3==1.26.14
Werkzeug==2.2.3
yarl==1.8.2